Moved UseCase into surfing.
